Configure repository
This commit configures the repository to use a clean/smudge filter for
making UTF-16 source and header files, created by Visual Studio,
readable when being checked in. The definition of the filter itself has
to be manually applied by appending the following snippet to
.git/config:

[filter "winutf16"]
        clean = iconv -f utf-16 -t utf-8
        smudge = iconv -f utf-8 -t utf-16
        required

It also configures the exclusion of the build folders and other files
created by the build process or Visual Studio.